META-2742 :Add Classification authz permission during partial AtlasGlossaryTerm update.

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/EntityGraphMapper.java
@@ -24,6 +24,9 @@ import org.apache.atlas.AtlasErrorCode;
 import org.apache.atlas.GraphTransactionInterceptor;
 import org.apache.atlas.RequestContext;
 import org.apache.atlas.annotation.GraphTransaction;
+import org.apache.atlas.authorize.AtlasAuthorizationUtils;
+import org.apache.atlas.authorize.AtlasEntityAccessRequest;
+import org.apache.atlas.authorize.AtlasPrivilege;
 import org.apache.atlas.exception.AtlasBaseException;
 import org.apache.atlas.exception.EntityNotFoundException;
 import org.apache.atlas.model.TimeBoundary;
@@ -2162,6 +2165,11 @@ public class EntityGraphMapper {
                     continue;
                 }
 
+                AtlasEntityHeader entityHeader =  entityRetriever.toAtlasEntityHeaderWithClassifications(entityVertex);
+                AtlasAuthorizationUtils.verifyAccess(new AtlasEntityAccessRequest(typeRegistry, AtlasPrivilege.ENTITY_ADD_CLASSIFICATION, entityHeader, classification),
+                        "add classification: guid=", guid, ", classification=", classification.getTypeName());
+
+
                 if (propagateTags == null) {
                     RequestContext reqContext = RequestContext.get();
 
@@ -2339,6 +2347,11 @@ public class EntityGraphMapper {
             throw new AtlasBaseException(AtlasErrorCode.INSTANCE_GUID_NOT_FOUND, entityGuid);
         }
 
+        AtlasEntityHeader entityHeader = entityRetriever.toAtlasEntityHeaderWithClassifications(entityVertex);
+        AtlasAuthorizationUtils.verifyAccess(new AtlasEntityAccessRequest(typeRegistry, AtlasPrivilege.ENTITY_REMOVE_CLASSIFICATION,
+                        entityHeader, new AtlasClassification(classificationName)),
+                "remove classification: guid=", entityGuid, ", classification=", classificationName);
+
         deleteDelegate.getHandler().deletePropagatedClassification(entityVertex, classificationName, associatedEntityGuid);
     }
 
@@ -2352,6 +2365,11 @@ public class EntityGraphMapper {
         if (entityVertex == null) {
             throw new AtlasBaseException(AtlasErrorCode.INSTANCE_GUID_NOT_FOUND, entityGuid);
         }
+
+        AtlasEntityHeader entityHeader = entityRetriever.toAtlasEntityHeaderWithClassifications(entityVertex);
+        AtlasAuthorizationUtils.verifyAccess(new AtlasEntityAccessRequest(typeRegistry, AtlasPrivilege.ENTITY_REMOVE_CLASSIFICATION,
+                        entityHeader, new AtlasClassification(classificationName)),
+                "remove classification: guid=", entityGuid, ", classification=", classificationName);
 
         AtlasPerfTracer perf = null;
